Ring&Vibrate Mode cleanup
This makes two entries and cleans up some code.

+++ configs/default_hooks/sxmo_hook_ring.sh
@@ -25,41 +25,42 @@ if ! sxmo_modemcall.sh list_active_calls \
 	exit
 fi
 
-case "$(cat "$XDG_CONFIG_HOME"/sxmo/.ringmode)" in
-	Mute|Vibrate)
-		;;
-	*)
-		# In order for this to work, you will need to install playerctl and run playerctld
-		# In order for this to work with mpv, you will need to install mpv-mdis.
-		sxmo_playerctl.sh pause_all
-		;;
-esac
+# RING & VIBRATE MODE (DEFAULT)
+if [ ! -f "$XDG_CONFIG_HOME"/sxmo/.noring ] && [ ! -f "$XDG_CONFIG_HOME"/sxmo/.novibrate ]; then
+	sxmo_log "RING AND VIBRATE"
 
-# Start the mpv ring until another hook kill it or the max SXMO_RINGTIME or SXMO_RINGNUMBER is reached
-case "$(cat "$XDG_CONFIG_HOME"/sxmo/.ringmode)" in
-	Mute)
-	;;
-	Vibrate)
-		for _ in $(seq 5); do
-			sxmo_vibrate 1500
-			sleep 0.5
-		done &
-		echo "$!" > "$XDG_RUNTIME_DIR/sxmo.ring.pid"
-		;;
-	Ring)
-		timeout "$SXMO_RINGTIME" mpv --no-resume-playback --quiet --no-video \
-			--loop="$SXMO_RINGNUMBER" "$SXMO_RINGTONE" >/dev/null &
-		echo "$!" > "$XDG_RUNTIME_DIR/sxmo.ring.pid"
-	;;
-	*) #Default ring and vibrate
-		timeout "$SXMO_RINGTIME" mpv --no-resume-playback --quiet --no-video \
-			--loop="$SXMO_RINGNUMBER" "$SXMO_RINGTONE" >/dev/null &
-		MPVID=$!
-		echo "$MPVID" > "$XDG_RUNTIME_DIR/sxmo.ring.pid"
-		# Vibrate while mpv is running
-		while kill -0 $MPVID; do
-				sxmo_vibrate 1500
-				sleep 0.5
-		done
-		;;
-esac
+	# In order for this to work, you will need to install playerctl and run playerctld
+	# In order for this to work with mpv, you will need to install mpv-mdis.
+	sxmo_playerctl.sh pause_all
+
+	timeout "$SXMO_RINGTIME" mpv --no-resume-playback --quiet --no-video \
+		--loop="$SXMO_RINGNUMBER" "$SXMO_RINGTONE" >/dev/null &
+	MPVID=$!
+	echo "$MPVID" > "$XDG_RUNTIME_DIR/sxmo.ring.pid"
+	# vibrate while mpv is running
+	while kill -0 $MPVID; do
+		sxmo_vibrate 1500
+		sleep 0.5
+	done
+
+# RING-ONLY MODE
+elif [ ! -f "$XDG_CONFIG_HOME"/sxmo/.noring ] && [ -f "$XDG_CONFIG_HOME"/sxmo/.novibrate ]; then
+	sxmo_log "RING ONLY"
+
+	# In order for this to work, you will need to install playerctl and run playerctld
+	# In order for this to work with mpv, you will need to install mpv-mdis.
+	sxmo_playerctl.sh pause_all
+
+	timeout "$SXMO_RINGTIME" mpv --no-resume-playback --quiet --no-video \
+		--loop="$SXMO_RINGNUMBER" "$SXMO_RINGTONE" >/dev/null &
+	echo "$!" > "$XDG_RUNTIME_DIR/sxmo.ring.pid"
+
+# VIBRATE-ONLY MODE
+elif [ ! -f "$XDG_CONFIG_HOME"/sxmo/.novibrate ] && [ -f "$XDG_CONFIG_HOME"/sxmo/.noring ]; then
+	smxo_log "VIBRATE ONLY"
+	for _ in $(seq 5); do
+		sxmo_vibrate 1500
+		sleep 0.5
+	done &
+	echo "$!" > "$XDG_RUNTIME_DIR/sxmo.ring.pid"
+fi
